srcs/requirements/bonus/dsbot/app/requirements.txt
discord.py==1.7.3
discord_py_slash_command==3.0.3
python-dotenv==1.0.0
